# Identify and address some performance hot spots