--- a/README.adoc
+++ b/README.adoc
@@ -453,6 +453,17 @@ For TLS connections, you can use the `openssl` `s_client` command:
 echo  '<13>1 2020-04-18T15:16:09.956153-07:00 coconut tyler - - [timeQuality tzKnown="1" isSynced="1" syncAccuracy="505061"] hello world' | openssl s_client -connect localhost:6514
 ----
 
+
+=== Profiling
+
+Profiling `hotdog` is best done on a Linux host with the `perf` tool, e.g.
+
+[source,bash]
+----
+RUST_LOG=info perf record --call-graph dwarf -- ./target/debug/hotdog -c ./hotdog.yml
+perf report -ng
+----
+
 == Similar Projects
 
 `hotdog` was originally motivated by challenges with