I don't have a great benchmark suite to compare different releases of :hotdog: to one another.

To improve performance I was mostly looking at `perf` results and comparing one run to another.


The closest comparison that I can run is between `master` and this branch with my `client.rb` script which sends 10k JSON events along.

*master*
```
time: /target/debug/hotdog  18.77s user 1.85s system 82% cpu 24.907 total 
time: ruby client.rb  0.31s user 0.04s system 2% cpu 16.642 total
```

*this branch*
```
time: ./target/debug/hotdog  12.56s user 1.24s system 34% cpu 40.219 total 
time: ruby client.rb  0.30s user 0.05s system 3% cpu 10.180 total
```

This is obviously lacking in any serious performance analysis rigor, but I feel pretty confident that the overall CPU usage was definitely improved with some of these changes.

This pull request closes #26